validacao no precadastro

+++ app/Http/Controllers/PeoplesController.php
@@ -10,6 +10,8 @@ use App\Models\Status;
 use App\Models\People;
 use App\Models\Config_system;
 use App\Models\Institution;
+use App\Models\People_Groups;
+use App\Models\Users_Account;
 use App\Http\Controllers\Input;
 
 class PeoplesController extends Controller
@@ -103,11 +105,18 @@ class PeoplesController extends Controller
 
     public function storeprecadastro(Request $request)
     {
+        $you = auth()->user();
 
-        //inserir no banco correto
+        //pegar tenant
         $value = $request->session()->get('key');
         Config::set('database.connections.tenant.schema', $value);
+       
+        //validar se tem
+        $validarprecadastro = People::where('user_id', $you->id);
 
+        if ($validarprecadastro->count() == 0)
+        {     
+        //inserir no banco correto
         $people = new People();
         $people->name          = $request->input('name');
         $people->email         = auth()->user()->email;
@@ -122,9 +131,15 @@ class PeoplesController extends Controller
         $people->is_verify       = 'false';
         $people->sex       = $request->input('sex');
         $people->note       = 'PRECADASTRO';
+        $people->user_id = $you->id;
         $people->save();
         $request->session()->flash("success", 'Cadastrado com sucesso, aguardar aprovação do administrador');
         return redirect()->route('account.index');
+        }
+        else{
+            session()->flash("info", "Você já possuiu vinculo, aguarde um administrador aprovar o seu acesso.");
+            return redirect()->back();
+        }
     }
 
 
@@ -208,15 +223,36 @@ class PeoplesController extends Controller
      */
     public function destroy($id)
     {
-        Config::set('database.connections.tenant.schema', session()->get('conexao'));
-        $people = people::find($id);
-        if ($people) {
-            $people->delete();
-        }
-        session()->flash("warning", "Sucessfully deleted people");
-        return redirect()->route('people.index');
-    }
+        $this->pegar_tenant();
 
+        $validargrupo = People_Groups::where('user_id', $id);
+        $validaracesso = Users_Account::where('user_id', $id)->with('accountlist');  
+
+        if ($validargrupo->count() > 1)
+        {
+            session()->flash("info", "Pessoa possui vinculo com grupos, precisa desassociar");
+            return redirect()->back();
+
+        }
+        if ($validaracesso->count() > 1)
+        {
+            session()->flash("info", "Pessoa possui acesso vinculado, precisa remover o acesso a conta");
+            return redirect()->back();
+
+        }
+        //@dump($validaracesso->count());
+
+        if ($validargrupo->count() == 0) {
+
+           $people = people::find($id);
+            if ($people) {
+            $people->delete();
+            }
+            session()->flash("warning", "Sucessfully deleted people");
+            return redirect()->route('people.index');
+            return redirect()->back();
+        }
+    }
 
     public function searchHistoric(Request $request, People $people)
     {
